+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 * Fix breaking pineapple crashing the game
+* Fix block items and drops
+* Disabled the Tropicraft dimension in prod